+++ tiff-ios/io/TIFFByteWriter.m
@@ -29,7 +29,7 @@
     [self.os close];
 }
 
--(NSData *) getData{
+-(NSData *) data{
     return [self.os propertyForKey:NSStreamDataWrittenToMemoryStreamKey];
 }